refactor: move interfaces
